Make filtering logic more elegant and less buggy
Abstract the filtering logic so it properly handles multiple filters, as
well as "de-selecting" the filter (choosing "Select a ___").

This paves the way for adding the turn around time filter.

Also add luxon for date parsing.

Another improvement might be to merge the data type with each item, so
that we won't have to hard code exceptions like price into the function.