Add cost bar graph
This includes the basic functionality for a static bar graph displaying
costs.

We are displaying very simple data, only one axis, but the nivo graph
library does not offer a way to hide either x or y axis, only bar
labels. While it is probably better visually to display *only* bar
labels, nivo does not support this at this time.

As a workaround, hide the labels, relying on the y axis for value
display, and hide the x axis via a brittle CSS selector; nivo also does
not add or allow adding identifiers to elements, so the selector used is
brittle because it is purely positional. If the source mark shipped from
the nivo library changes, this selector could break.