Add top-level start/stop script
This PR aggregates both the server and browser default startup commands
into one single command. It also sends the tasks into the background,
saving the output to log files. The process IDs are saved, so they can
be stopped on demand.

This depends on #40 being complete before it will work, since the
interactive element of having to say "Yes" to a new port on the browser
stops that from starting.

Fixes #41