Add top-level start/stop script
This PR aggregates both the server and browser default startup commands
into one single command.

It uses pm2 to manage the processes, so an `npm install` is required in
the top level directory. I tried to manage the processes with standard
unix jobs, but the react server does not function properly when the job
is suspended.

I also added a top-level install command that installs modules in server
and browser as well, so doing a `npm install` at the top-level should be
all you need to update now.

I also updated the lock files for browser and server.

Fixes #41