Adjust charts for small screens
On small screens, the charts have excess white space around them,
causing the data to unnecessarily overlap. The nivo library does not
automatically reformat it's margins based on the data, leaving it as a
manual process; if you don't provide enough, the data will be cut off.

As a solution, abbreviate the numbers in the cost chart to the thousands
place and round all values. Provide the full number on hover. This
allows us to reduce the amount of margin needed, preventing data overlap
on most common screens.