Add funnel chart dependency
The funnel chart dependency tree is out of sync with other nivo
dependencies, throwing an error when trying to install it. However, it
seems like the upstream version numbers just weren't bumped, and the
package works fine in my tests. I force installed it and was able to get
a dummy funnel chart in the app.

re #69