Sandbox Graphs and Componentization (#75)
* Add funnel chart dependency

The funnel chart dependency tree is out of sync with other nivo
dependencies, throwing an error when trying to install it. However, it
seems like the upstream version numbers just weren't bumped, and the
package works fine in my tests. I force installed it and was able to get
a dummy funnel chart in the app.

re #69

* Reorganzing the functions

* common data pull

* adding react router, adding new sandbox page

* added treepmap to sandbox page

* Created helper classes for turnaround times and fee ranges

* timeline graph component

* treemap component

* filter using turnaround times

* remove comments

* fixed graph sizing

* funnel and pie graphs

* fucking sankey y'alllllllklklllllllllllll

* adding node_modules to git ignore

* adding fee bubble graph

* Fee bubble is working

* currency tooltip

* fixing up some labels and tooltips

* tweak my bubblessss

Co-authored-by: Chris DeLuca <code@chrisdeluca.me>